pyproject.toml
[project]
name = "project-crew"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "abdul rehman", email = "157266271+moiz14553@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.140.0",
    "crewai-tools>=0.49.0",
    "python-dotenv",
    "litellm",
]

[project.scripts]
projectflow = "project_crew.main:kickoff"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
